Move tarball creation to its own class
