+++ app/jobs/exercise_verifier_job.rb
@@ -1,10 +1,9 @@
 # frozen_string_literal: true
 
 class ExerciseVerifierJob < ApplicationJob
-  require 'test_generator'
   require 'main_class_generator'
+  require 'test_generator'
   require 'rest-client'
-  require 'submission_status_channel'
   require 'timeout'
   queue_as :default
 
@@ -17,43 +16,7 @@ class ExerciseVerifierJob < ApplicationJob
       Rails.logger.warn('ExerciseVerifierJob called with a string! Silently aborting...')
       return
     end
-
-    if !exercise_modified?(exercise)
-      exercise.error_messages.push(header: 'Tapahtui virhe: Muokkaamaton tehtävä lähetettiin uudelleen')
-      exercise.error!
-      MessageBroadcasterJob.perform_now(exercise)
-    else
-      send_exercise_to_sandbox(exercise)
-    end
-  end
-
-  def exercise_modified?(exercise)
-    if Dir.exist?(assignment_target_path(exercise)) && Dir.exist?(assignment_target_path(exercise).join("exercise_#{exercise.id}"))
-      if directory_includes_file(exercise, 'ModelSolution') || directory_includes_file(exercise, 'Template')
-        return false
-      end
-    end
-    true
-  end
-
-  def directory_includes_file(exercise, package_type)
-    Dir.entries(assignment_target_path(exercise).join("exercise_#{exercise.id}")).include?("#{package_type}_#{exercise.id}.#{exercise.versions.last.id}.zip")
-  end
-
-  def create_tar_files(exercise)
-    exercise.create_submission
-
-    `cd #{tmp_submission_target_path(exercise).join('model').to_s} && tar -cpf #{packages_target_path.join("ModelSolutionPackage_#{exercise.id}.tar").to_s} *`
-    `cd #{tmp_submission_target_path(exercise).join('template').to_s} && tar -cpf #{packages_target_path.join("TemplatePackage_#{exercise.id}.tar").to_s} *`
-  end
-
-  def send_exercise_to_sandbox(exercise)
-    create_tar_files(exercise)
-
-    # Send template to sandbox
     send_package_to_sandbox(exercise, 'TEMPLATE', "TemplatePackage_#{exercise.id}.tar")
-
-    # Send model solution to sandbox
     send_package_to_sandbox(exercise, 'MODEL', "ModelSolutionPackage_#{exercise.id}.tar")
   end
 
@@ -93,15 +56,7 @@ class ExerciseVerifierJob < ApplicationJob
     "#{ENV['BASE_URL']}/api/v0/exercises/#{exercise.id}/results"
   end
 
-  def assignment_target_path(exercise)
-    Rails.root.join('submission_generation', 'packages', "assignment_#{exercise.assignment.id}")
-  end
-
   def packages_target_path
     Rails.root.join('submission_generation', 'packages')
-  end
-
-  def tmp_submission_target_path(exercise)
-    Rails.root.join('submission_generation', 'tmp', "Submission_#{exercise.id}")
   end
 end
